Refactor: Merge assembly_model into assembler and fix regression test location
- Merged `src/semantic/assembly_model.rs` into `src/semantic/assembler.rs` to reduce file count and indirection.
- Deleted `src/semantic/assembly_model.rs`.
- Moved `test_dropped_operator` from `tests/regression_operator_drop.rs` to `src/semantic/expressions.rs` as it tests internal API.
- Deleted `tests/regression_operator_drop.rs`.
- Updated `.jules/razor.md` with the reduction.
- Verified with `cargo test`.
