+++ src/semantic/mod.rs
@@ -34,7 +34,6 @@
 pub mod assembler;
 #[cfg(test)]
 mod assembler_tests;
-pub(crate) mod assembly_model;
 pub(crate) mod conversion;
 pub mod expressions;
 pub(crate) mod model;
@@ -44,9 +43,8 @@ pub(crate) mod statements;
 mod types;
 
 pub use crate::morphology::{DisambiguationContext, analyze_article, disambiguate, resolve_best};
-pub(crate) use assembler::Assembler;
 pub use assembler::AssemblyError;
-pub(crate) use assembly_model::{AssembledStatement, Constituent, Literal};
+pub(crate) use assembler::{AssembledStatement, Assembler, Constituent, Literal};
 pub use model::*;
 pub use resolver::*;
 pub use types::*;
